requirements.txt
selenium>=4.9.0
pyyaml>=6.0.1
requests>=2.28.0
webdriver-manager>=3.8.0